requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
pydantic
psycopg2-binary
